Glue for Lua is working.
This implements `sol` and `eol` for start-of-line() and end-of-line()
respectively.

The invokation of these functions is handled at the Lua layer.  In
an ideal world we'd translate the characters from input to KEY_HOME,etc
in Lua too, but this is a sleazy hack.

We could also do prefix-matching to allow binding things to Ctrl-x
Ctrl-s too if we got fancy.