Implement at(), mark() and allow multi-line cut.
This update introduces the notion of a "mark".  A mark
is a (global) recorded cursor position.

You can set a mark with:

* ctrl+space

Once a mark is set you can DELETE the region between the
point and the mark with Ctrl-w and paste it back via Ctrl-u.

This implements emacs-like copy/paste, however the biggest
omission is that you cannot see the mark.

Shame.

NOTE: Reordered the kilo.lua file too.