Wire up undo-stack.
This turned out to be about as easy as expected; for every
action the user takes we need to merely record the opposite
in our stack.

However there are complications:

* Anything that warps the cursor will break the undo-stack.

This meant that the `eol` / `sol` primitives had to be rewritten
to count the number of LEFT/RIGHT-arrow movements, such that they
could be undone in turn.

It also means that the oment the user hits Ctrl-f (find) the
undo stack will break, because the cursor is warped without
the use of the undo-primitives.

This updates #14.