Only highlight tokens, via regexp, at offset=0.
Given this line:

     if end if

We step forwards character-by-character looking for a match against
each of our tokens.  For the regular expression case we might search
for 'end' and that would match - even if we're searching for the
start of the string.

So we need to ensure that the match is at the current X-position we're
looking for.  WIthout that we'll get screwy highlighting.