The enemy of good is perfect.
Commit the work I've done on my "new" editor, which supports
UTF-8 and is still primarily lua-driven.

The `README.md` file should explain a lot, the `TODO.md` file
should explain more.

Questions?  Report a bug.

This is explicitly not documented just yet, but will be in the
future.

This closes #49.