Added paranoid compiler options.
This caused immediate crashes, due to an off-by-one in the
character conversion, and some issues with the use of "delete"
vs "delete[]".

Finally we were allocating editorState as a struct, but using
the member-variable `std::vector` which thus wasn't initialized
by malloc.

In short we had bugs, now they're fixed.

This closes #58.