Implement setting the mark and drawing the same.
I'm 99% certain this is correct, on my third attempted at handling
setting the mark and drawing the selected region.

In brief:

* We have a buffer containing rows.
* The rows contain characters.

We map from X,Y to the starting offset of the text
such that we conver (0,0) to text[0], and (0,10) to
text(10).

Then when drawing the text we keep track of the characters
we've encountered and if the current character is within the
appropriate range we configure reverse video.

This updates #71 to *highlight* the selection, but does not
yet allow copying or cutting.

The plan is to clone the approach, even if inefficient, to
allow selection-copying and cutting.