Rework drawing of screen.
This updates the drawing of the characters in a given row
to account for the characters which are to the left of the current
display-offset, and to the right.

In short it bumps things around such that we keep count of the
characters we've drawn, which allows the selection to be drawn
correctly even if we're handling long-lines and horizontal
scrolling.

This closes #73.