Update syntax-highlighting in the background.
This is a significant change which drops the `on_syntax_highlight`
callback.  Instead the same operations happen, but only in the
`on_idle()` callback.

Profiling showed that the reason the editor was laggy was that
the constant calling to `on_syntax_highlight()` was taking too
long.

Even though caching the modification state of the buffer was working
the fact is that when you open a buffer and type characters you
expect it to update immediately, the modification-time would get
bumped on every keypress triggering a re-render, and essentially
making the cache pointless.

We can improve further here, because the on_idle() function should
also be able to cache the current state of the buffer- but because
it is only called when things are idle it actually doesn't matter.

This closes #78.