Allow searching fowards to work, even on a match.
Previously if hte point was at the start of a match then it
would be found as a result.  Meaning if there were N matches
you'd only find the first - then never move.

I've updated the search() function to return `true` if a hit
was found.  If that is the case AND the cursor didn't change then
we know we were under a match at the starting point, so we move
once to the right, and try again.

This closes #93.